requirements.txt
Django
djangorestframework 
django-ninja
channels
channels-redis
redis
celery
requests
